pull player into routes/player
